More folder renaming for better terminology; add some documentation
