+++ frontend/src/components/widgets/floating-menus/ColorPicker.vue
@@ -118,9 +118,9 @@
 <script lang="ts">
 import { defineComponent, PropType } from "vue";
 
-import { RGBA } from "@/interop/messages";
-import { hsvaToRgba, rgbaToHsva } from "@/utilities/color";
-import { clamp } from "@/utilities/math";
+import { hsvaToRgba, rgbaToHsva } from "@/utility-functions/color";
+import { clamp } from "@/utility-functions/math";
+import { RGBA } from "@/wasm-communication/messages";
 
 import LayoutCol from "@/components/layout/LayoutCol.vue";
 import LayoutRow from "@/components/layout/LayoutRow.vue";
