+++ frontend/src/state-providers/portfolio.ts
@@ -1,9 +1,9 @@
 /* eslint-disable max-classes-per-file */
 import { reactive, readonly } from "vue";
 
-import { Editor } from "@/interop/editor";
-import { TriggerFileDownload, TriggerRasterDownload, FrontendDocumentDetails, TriggerFileUpload, UpdateActiveDocument, UpdateOpenDocumentsList } from "@/interop/messages";
-import { download, downloadBlob, upload } from "@/utilities/files";
+import { download, downloadBlob, upload } from "@/utility-functions/files";
+import { Editor } from "@/wasm-communication/editor";
+import { TriggerFileDownload, TriggerRasterDownload, FrontendDocumentDetails, TriggerFileUpload, UpdateActiveDocument, UpdateOpenDocumentsList } from "@/wasm-communication/messages";
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export function createPortfolioState(editor: Editor) {
