+++ frontend/src/state-providers/workspace.ts
@@ -1,8 +1,8 @@
 /* eslint-disable max-classes-per-file */
 import { reactive, readonly } from "vue";
 
-import { Editor } from "@/interop/editor";
-import { UpdateNodeGraphVisibility } from "@/interop/messages";
+import { Editor } from "@/wasm-communication/editor";
+import { UpdateNodeGraphVisibility } from "@/wasm-communication/messages";
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export function createWorkspaceState(editor: Editor) {
