+++ frontend/src/utility-functions/icons.ts
