+++ frontend/src/utility-functions/keyboard-entry.ts
