+++ frontend/src/utility-functions/math.ts
