+++ frontend/src/utility-functions/strip-indents.ts
