+++ frontend/src/wasm-communication/subscription-router.ts
@@ -1,7 +1,7 @@
 import { plainToInstance } from "class-transformer";
 
-import type { WasmEditorInstance, WasmRawInstance } from "@/interop/editor";
-import { JsMessageType, messageMakers, JsMessage } from "@/interop/messages";
+import type { WasmEditorInstance, WasmRawInstance } from "@/wasm-communication/editor";
+import { JsMessageType, messageMakers, JsMessage } from "@/wasm-communication/messages";
 
 type JsMessageCallback<T extends JsMessage> = (messageData: T) => void;
 type JsMessageCallbackMap = {
