+++ frontend/wasm/src/lib.rs
@@ -19,8 +19,10 @@ thread_local! {
 /// Initialize the backend
 #[wasm_bindgen(start)]
 pub fn init() {
+	// Set up the panic hook
 	panic::set_hook(Box::new(panic_hook));
 
+	// Set up the logger with a default level of debug
 	log::set_logger(&LOGGER).expect("Failed to set logger");
 	log::set_max_level(log::LevelFilter::Debug);
 }
